backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-jose[cryptography]==3.3.0
google-generativeai==0.8.3
supabase==2.9.0
python-dotenv==1.0.1
elevenlabs==1.9.0
serpapi==0.1.5
slowapi==0.1.9
